Moves ctrl-word-movement to hammerspoon, adds fn+arrow delete
My ctrl-word-modement remaps are now available system-wide. Without
any complex hammerspoon code for checking if Emacs was the frontmost
application. To make this work I just needed to verify that Emacs
understood all the keys hammerspoon was sending. Which was trivial.So
this is a much simpler implimentation.

The one downside is that I can no longer use ^u and ^w in Terminal
because hammerspoon sends cmd+delete and alt+delete instead. But
that's a minor annoyance since I spend most of my days in Emacs now.

Also turns fn+left/right into delete/forward-delete, and allows for
the correct alt and cmd modifiers. This is really elegant. If you're
moving/selecting with the arrow keys already you can just add the 'fn'
modifier, just like you would with 'shift' to delete text. Beautifully
simple.