Fixes the narrowing/completion setup. Working now!
Finally got a bunch of stuff that just didn't work working. For one thing,
this set of packages is under very active development, so it pays to really be
up to date. At this moment in time nothing in my setup is broken.

I /finally/ have a working setup for using Selectrum in combination with
Embark. But I was just lucky. There were significant upgrades in the last few
days and I stumbled on a github issue that contained a function and hook that
makes Embark work exactly the way I want it. So that's nice.