Font Settings Updates
Largely sparked by an update to modus-themes 1.1 and switching to using Inter
as my variable pitch font.

My composition mode is too big a hammer with which I get Eww and Elfeed
looking correct, with this new set of fonts anyway. It's much better to let
each stick with the defaults and tweak the minimum needed for it to look
better.