Moves package config later in the init file
This allows everything above to work correctly when installing on a machine
that has no packages installed yet, for example, all my keybindings and
functions now work with no packages installed. If the package config is at the
top then Emacs will stop loading the init file when it encounters something
from a non-installed package.