Major re-thinking of keybindings and transients
Mostly this was done our of a desire to stop relying on the super key for my
most commonly used commands, and reduce the number of transients I have
bindings for. An overall simplification that, as a bonus, will allow my
experience on other platforms (Windows) to be more consistent.